Main Gui for Autotilting
Attaching the main script which uses the auxiliary code to create an interface capable of reorienting the sample to the ideal conditions for high resolution imaging in STEM. This requires a user to currently click where the location of the Kikuchi Zone Axis and Ronchigram centre is, and for the relevant calibrations to have been completed for the given set-up. Note once the user has calibrated a specific centre, the information is stored and can be selected via a dropdown box. No further calibration is needed.